final_rul_model/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.4.2
joblib>=1.3.0
kafka-python>=2.0.2
python-dateutil>=2.8.2
requests>=2.31.0